chore: photo_metadata のテストを別ファイルに切り出し clippy を通す
get_stored_capture_times のテストが加わって photo_metadata.rs が行数上限を超え、
編集できなくなっていた。テストモジュールを photo_metadata_tests.rs へ移す。
directory.rs が directory_tests.rs を #[path] で読み込むのと同じ形。中身は
インデントを1段戻した以外そのまま。

あわせて cargo clippy --all-targets -- -D warnings (pre-push フックが走らせる)
が落ちていた unnecessary_get_then_check を直す。--bin photoclove だけでは
テストコードが対象外で気づけなかった。
